Add a FlagSpec registry for node and edge flags (#280)
Node and edge flags were a fixed catalog of compile-time bits with no
name table, so a plugin could only emit the built-ins and an external
reader of the graph file couldn't map a bit back to a name. Introduce a
single FlagSpec mechanism with two registration paths: the engine
registers its engine/* flags with explicit masks (hot-path reads still
constant-fold), and plugins declare flags via the new
declare_node_flags()/declare_edge_flags() hooks, with the host allocating
each a bit above the engine masks in registration order (deterministic).
The engine/ namespace is reserved, identical specs registered by two
plugins share one bit (pytest and unittest now share test/testcase), and
a conflicting re-declaration or exhausting the bit width fails loudly.

Both registries are serialized into the .dcg file (FORMAT_VERSION 1->2,
old files rejected) so an external explorer can decode any bit by name.
Node flags stay u32; edge flags narrow to u8.

Roster surgery: drop the dead SHADOWED and EXPORTED flags; move OVERLOAD
off the flag space into an internal build-time is_overload field on the
node payload (not serialized, not Python-visible); move TESTCASE to the
shared test/testcase plugin flag; add an engine dead_branch node flag.
The hand-maintained KEEPALIVE_DEFAULT constant is replaced by the
registry-derived default_seed_mask() (OR of seed && default_on bits), so
test/testcase is kept alive by default iff a test plugin is registered.

PLUGIN_API_EPOCH bumped 1->2 (the curated plugin_api surface changed).

+++ runtime/src/helpers.rs
@@ -896,8 +896,8 @@ pub(crate) const NODE_FLAGS_NOQA_PIN: u32 = NodeFlags::NOQA;
 /// `NodeFlags::ENTRYPOINT` (which would force every reader to chase the
 /// `pyclass` macro to decide whether it's a runtime lookup or a const).
 pub(crate) const NODE_FLAG_ENTRYPOINT: u32 = NodeFlags::ENTRYPOINT;
-pub(crate) const EDGE_FLAG_DEAD_BRANCH: u32 = EdgeFlags::DEAD_BRANCH;
-pub(crate) const EDGE_FLAG_DYNAMIC_IMPORT: u32 = EdgeFlags::DYNAMIC_IMPORT;
+pub(crate) const EDGE_FLAG_DEAD_BRANCH: u8 = EdgeFlags::DEAD_BRANCH;
+pub(crate) const EDGE_FLAG_DYNAMIC_IMPORT: u8 = EdgeFlags::DYNAMIC_IMPORT;
 
 /// Collect the source ranges of every statically-dead statement in
 /// `parsed` — bodies of `if False` / `else` of `if True` / `while False`
